move the NUTS proposal to the NUTS kernel file
